Move the LoggedRule trait closer to Rule
I think this doesn't actually need to be a trait, that's needlessly
OOPy.
