+++ src/achievement/mod.rs
@@ -5,5 +5,5 @@ mod process_rules;
 #[cfg(test)]
 mod test_process_rules;
 
-pub use achievement::{Achievement, Rule};
+pub use achievement::{Achievement, LoggedRule, Rule};
 pub use process_rules::{grant, grant_with_rules};
